Trunk check-in of the Generic Word Space.
Additional features:
- vector support for sparse integer vector views
- bug fixes among the various vectors.  
- moved BasisMapping to its own package and added a default implementation


A  + test/edu/ucla/sspace/vector/ViewDoubleAsDoubleSparseVectorTests.java
A  + test/edu/ucla/sspace/vector/SparseIntegerVectorViewTests.java

- Added unit tests for the view sparse view vectors to address certain bugs and
  performance issues

MM   src/edu/ucla/sspace/tools/SemanticSpaceExplorer.java

- Added new instruction for getting the description of dimensions

- Added new instruction for getting statistics on the currently loaded semantic
  space

A  + src/edu/ucla/sspace/basis
A  + src/edu/ucla/sspace/basis/GenericBasisMapping.java
A  + src/edu/ucla/sspace/basis/BasisMapping.java
M    src/edu/ucla/sspace/dv/DependencyPathBasisMapping.java
D    src/edu/ucla/sspace/dv/BasisMapping.java

- Moved the basis mapping interface to its own package for now, as it's not
  really specific to the dv package

M    src/edu/ucla/sspace/vector/IntegerVectorView.java
M    src/edu/ucla/sspace/vector/ViewDoubleAsDoubleSparseVector.java
MM   src/edu/ucla/sspace/vector/DoubleVectorView.java
M    src/edu/ucla/sspace/vector/VectorView.java

- Fixed bugs in getNonZeroIndices() and added a performance improvement

- Fixed race condition and bug in the magnitude()

MM   src/edu/ucla/sspace/vector/Vectors.java
A  + src/edu/ucla/sspace/vector/SparseIntegerVectorView.java

- Added a subview method for the SparseIntegerVector

MM   src/edu/ucla/sspace/common/Similarity.java

- Minor performance enhancement to cosineSimilarity for IntegerVector instances
  to prefer int arithmetic to double.

M    src/edu/ucla/sspace/common/SemanticSpaceIO.java

- Fixed bug in loading a serializable semantic space.

A  + src/edu/ucla/sspace/gws
A  + src/edu/ucla/sspace/gws/GenericWordSpace.java

- Added a generic co-occurrence word space, with optional word-position counting

A  + src/edu/ucla/sspace/gws/WordOrderBasisMapping.java
A  + src/edu/ucla/sspace/gws/WordBasisMapping.java

- Added specialized basis mappings for the GWS dimension descriptions

A  + src/edu/ucla/sspace/mains/GwsMain.java

- Added a main() runnable for GWS

MM   build.xml

- Added a gws-jar target

+++ trunk/sspace/src/edu/ucla/sspace/common/Similarity.java
@@ -375,7 +375,7 @@ public class Similarity {
     public static double cosineSimilarity(IntegerVector a, IntegerVector b) {
         check(a,b);
 
-        double dotProduct = 0.0;
+        int dotProduct = 0;
         double aMagnitude = a.magnitude();
         double bMagnitude = b.magnitude();
 
@@ -425,15 +425,15 @@ public class Similarity {
             // misses would be less of a performance hit.
             if (nzA.length < nzB.length) {
                 for (int nz : nzA) {
-                    double aValue = a.get(nz);
-                    double bValue = b.get(nz);
+                    int aValue = a.get(nz);
+                    int bValue = b.get(nz);
                     dotProduct += aValue * bValue;
                 }
             }
             else {
                 for (int nz : nzB) {
-                    double aValue = a.get(nz);
-                    double bValue = b.get(nz);
+                    int aValue = a.get(nz);
+                    int bValue = b.get(nz);
                     dotProduct += aValue * bValue;
                 }
             }
@@ -442,8 +442,8 @@ public class Similarity {
         // Otherwise, just assume both are dense and compute the full amount
         else {
             for (int i = 0; i < b.length(); i++) {
-                double aValue = a.get(i);
-                double bValue = b.get(i);
+                int aValue = a.get(i);
+                int bValue = b.get(i);
                 dotProduct += aValue * bValue;
             }
         }
